refactor with CookieStorage
